+++ Code/Bridge144/lib/Parameters/parameterManager_YAML.cpp
@@ -542,7 +542,7 @@ void ParameterManager_YAML::read_params(const std::string& params_file, Paramete
     }
 
     fin.seekg(0, std::ios::end);
-    filesize = fin.tellg();
+    filesize = (int)fin.tellg();
     fin.seekg(0, std::ios::beg);
 
     int padding = 8 - (filesize % 8);
